requirements.txt
beautifulsoup4==4.12.0
crispy-bootstrap5==0.7
django-debug-toolbar==3.8.1
django-extensions==3.2.1
django-htmx==1.13.0
django-phac_aspc-helpers==0.4.3
django-simple-history==3.2.0
django-versionator==0.0.1b0
Django==4.1.9
Faker==15.1.1
freezegun==1.2.2
gunicorn==20.1.0
graphene==3.1.1
djangorestframework==3.14.0
graphene-django==3.0.0
graphql-core==3.2.3
ipython==8.10.0
Jinja2==3.1.2
openpyxl==3.0.10
promise==2.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.2
whitenoise==6.4.0
